Move TokenClients to token endpoint
