+++ src/contracts/latest/index.ts
@@ -4,5 +4,3 @@
 export type { ColonyExtension } from "./ColonyExtension";
 export type { IBasicMetaTransaction } from "./IBasicMetaTransaction";
 export type { IColonyNetwork } from "./IColonyNetwork";
-export type { TokenAuthority } from "./TokenAuthority";
-export type { TokenLocking } from "./TokenLocking";
