+++ src/extras/index.ts
@@ -39,7 +39,7 @@ export {
   LogSetOwnerEventObject,
 } from '../contracts/latest/ColonyExtension';
 
-export * from '../contracts/latest/TokenLocking';
+export * from '../tokens/contracts/latest/TokenLocking';
 
 export * from '../contracts/__dynamic__/CoinMachineEvents';
 export * from '../contracts/__dynamic__/IColonyEvents';
