+++ src/types.ts
@@ -6,8 +6,8 @@ import type { ColonyRole } from './constants';
 import { AnyColonyClient } from './clients/Core/exports';
 import { ExtensionClient } from './clients/Extensions/exports';
 import { ColonyNetworkClient } from './clients/ColonyNetworkClient';
-import { TokenClient as TokenContractClient } from './clients/TokenClient';
-import { TokenLockingClient as TokenLockingContractClient } from './clients/TokenLockingClient';
+import { TokenClient as TokenContractClient } from './tokens/clients/TokenClient';
+import { TokenLockingClient as TokenLockingContractClient } from './tokens/clients/TokenLockingClient';
 
 import {
   CoinMachineEvents,
